requirements.txt
sqlalchemy>=1.4
psycopg2-binary
python-dotenv
pandas
openpyxl
matplotlib
tabulate
colorama